refactor: courses controllers refactored
