+++ backend/src/controllers/courses/getCourse/utils/mappers.ts
@@ -1,0 +1,8 @@
+import { ICourseWithStatus } from 'interfaces/ICourses/IQueryCourses';
+
+const mapCourse = (course: ICourseWithStatus, similarCourses: never[]): ICourseWithStatus => ({
+  ...course,
+  similarCourses,
+});
+
+export { mapCourse };
