+++ backend/src/controllers/courses/index.ts
@@ -1,0 +1,4 @@
+export { default as getAllCourses } from './getAllCourses/getAllCourses';
+export { default as getCourseById } from './getCourse/getCourse';
+export { default as getCoursesMap } from './getCoursesMap/getCoursesMap';
+export { default as getMaterials } from './getMaterials/getMaterials';
